(#117) (#45) tidy tests
